Course Ownership: Only the owner of the bootcamp can add courses to it.Only the creator of the course can update and delete it.

+++ controllers/courses.js
@@ -57,6 +57,7 @@ exports.getCourse = async (req, res, next) => {
 exports.addCourse = async (req, res, next) => {
   try {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
     if (!bootcamp) {
       return next(
@@ -64,6 +65,15 @@ exports.addCourse = async (req, res, next) => {
           `No bootcamp with the id of ${req.params.bootcampId}`
         ),
         404
+      );
+    }
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() != req.user.id && req.user.role != "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+          401
+        )
       );
     }
     const course = await Course.create(req.body);
@@ -87,6 +97,15 @@ exports.updateCourse = async (req, res, next) => {
       return next(
         new ErrorResponse(`No course with the id of ${req.params.id}`),
         404
+      );
+    }
+    // Make sure user is course owner
+    if (course.user.toString() != req.user.id && req.user.role != "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to update course${course._id}`,
+          401
+        )
       );
     }
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -113,6 +132,15 @@ exports.deleteCourse = async (req, res, next) => {
       return next(
         new ErrorResponse(`Course not found with id of ${req.params.id}`, 404)
       );
+    // Make sure user is course owner
+    if (course.user.toString() != req.user.id && req.user.role != "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.user.id} is not authorized to delete course${course._id}`,
+          401
+        )
+      );
+    }
     course = await Course.findByIdAndDelete(req.params.id);
     res.status(200).json({ success: true, data: course });
   } catch (err) {
